发现bug,修复一下

+++ fileshen/src/main/java/tan/wei/feng/controller/FileController.java
@@ -23,7 +23,6 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.MalformedJwtException;
 import tan.wei.feng.entity.Result;
 import tan.wei.feng.entity.StatusCode;
 import tan.wei.feng.service.create.FileSaveService;
@@ -65,18 +64,11 @@ public class FileController {
 	 */
 	@GetMapping(value = "/fileList/{fileType}")
 	public Result fileUrlList(@PathVariable String fileType){
-		
-		try {
-			Claims claims = (Claims) request.getAttribute("user_claims");
-			logger.info(claims.getId());
-			if(claims.getId()==null || "".equals(claims.getId().trim())){
-				return new Result(StatusCode.ERROR);
-			}
+		Claims claims = (Claims) request.getAttribute("user_claims");
+		if(claims != null && !"".equals(claims.getId().trim())){
 			return new Result(StatusCode.OK,"查询成功",fileFindService.findByfileid(claims.getId(),fileType));
-		}catch (Exception e) {
-			logger.info(e.getMessage());
-			return new Result(StatusCode.ERROR);
 		}
+		return new Result(StatusCode.ERROR,"请登录在查询");
 		
 	}
 	
@@ -87,13 +79,10 @@ public class FileController {
 	@GetMapping(value = "/filePageList/{fileType}")
 	public Result fetchList(@RequestParam Map<String,String> map,@PathVariable String fileType){
 		Claims claims=(Claims) request.getAttribute("user_claims");
-		if(claims == null){
-			return new Result(StatusCode.ERROR,"请登录在查询");
-		}
-		if(claims.getId() != null && !"".equals(claims.getId().trim())){
+		if(claims != null && !"".equals(claims.getId().trim())){
 			return new Result(StatusCode.OK,"查询成功",fileFindService.findByfileidPage(claims.getId(),fileType,map));
 		}
-		return new Result(StatusCode.ERROR,"请上传文件");
+		return new Result(StatusCode.ERROR,"请登录");
 	}
 	
 	
@@ -105,16 +94,13 @@ public class FileController {
 	 * @throws IOException
 	 */
 	@GetMapping(value = "/fileDownload/{filecode}")
-	public ResponseEntity<?> fileDownload(@PathVariable String filecode) throws IOException {  
+	public ResponseEntity<byte[]> fileDownload(@PathVariable String filecode) throws IOException {  
 		Claims claims=(Claims) request.getAttribute("user_claims");
-		logger.info(claims.getId());
-		if(claims==null || "".equals(claims.getId())){
-			Result result = new Result("请登录在下载"); 
-			// 也可以放置页面,图片等
-			return new ResponseEntity<Result>(result, HttpStatus.UNAUTHORIZED);
-		} else {
+		if(claims != null && !"".equals(claims.getId().trim())){
+			logger.info(claims.getId());
 	        return fileDownService.fileDownload(filecode);
 		}
+		return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 	}
 	
 	
@@ -128,25 +114,22 @@ public class FileController {
 	@PostMapping(value = "/uploadFile")
 	public Result fileSave(@RequestParam("file") MultipartFile file){
 		Claims claims=(Claims) request.getAttribute("user_claims");
-		if(claims==null || "".equals(claims.getId())){
-			return new Result(StatusCode.ERROR,"请登录在上传");
-		}  
-		if(!file.isEmpty()){
+		if(!file.isEmpty() && claims != null && !"".equals(claims.getId().trim())) {
 			// 递归次数
 			int ia = 3;
 			String filePath = FileDirUtils.dirFile(datapath,file.getOriginalFilename(),ia) ;
 			//转存文件
 	        try {
 	            file.transferTo(new File(filePath));
+	            if(fileSaveService.addFile(filePath,claims.getId())) {
+					return new Result(StatusCode.OK,"上传成功");
+				} 
 	        } catch (Exception e) {
 	        	File filea = new File(filePath);
 	        	filea.delete();
-	            return new Result(StatusCode.ERROR,e.getMessage());
+	            return new Result(StatusCode.ERROR,"上传失败重新上传"+e.getMessage());
 	        }
-	        if(fileSaveService.addFile(filePath,claims.getId())) {
-				return new Result(StatusCode.OK,"上传成功");
-			} 
-	     }
+		}
 		return new Result(StatusCode.ERROR,"上传失败重新上传");
 	}
 		
@@ -158,15 +141,11 @@ public class FileController {
 	@DeleteMapping(value = "/delFile/{fileId}")
 	public  Result delFile(@PathVariable String fileId) {
 		Claims claims=(Claims) request.getAttribute("user_claims");
-		if(claims==null || "".equals(claims.getId())){
-			return new Result(StatusCode.ERROR,"请登录在下载");
-		}else {
-			if(fileDeleteService.delfile(Long.parseLong(fileId))) {
-				return new Result(StatusCode.OK,"删除成功");
-			} else {
-				return new Result(StatusCode.ERROR,"删除失败");
-			}
+		if(claims != null && !"".equals(claims.getId().trim()) && fileDeleteService.delfile(Long.parseLong(fileId))){
+			return new Result(StatusCode.OK,"删除成功");
 		}
+		return new Result(StatusCode.ERROR,"删除失败");
+		
 	}
 
 }
